refactor lambda utils
